Fixed the cart view losing size information
The cart was displaying two products that were the same, and but
differed in size, as the same product line. For example, if I bought
some pants, one large and one small, it would show in the cart as 1
item, quantity 2, both large, instead of two separate items. This has
been fixed.